models/code/requirements.txt
diffusers
ipython
json5
peft==0.13.
pillow
safetensors==0.4.5
tokenizers
torch
torchaudio
torchvision
tqdm
transformers==4.46.3
accelerate